moved a bunch of components around
