+++ src/app/routes/boom/boom.component.ts
