+++ src/app/routes/curriculum-editor/create-version/create-version.component.ts
