+++ src/app/routes/curriculum-editor/curriculum-editor.component.css
