+++ src/app/routes/curriculum-editor/curriculum-editor.component.spec.ts
